Revision:
DONE:
	- the sizes of stack and constant pool are both limited to 255
	  entries;

TODO:
	- ensure that current implementation matches with the one
	  discussed in 'Crafting Interpreters' till the chapter 21
"Global Variables"

+++ src/object.h
@@ -60,7 +60,7 @@ struct Object {
 */
 typedef struct {
 	Object		object;
-	uint32_t	arity;		/* the number of params. */
+	uint8_t		arity;		/* the number of params. */
 	Bytecode	bytecode;	/* function's own bytecode. */
 	ObjString*	name;		/* Function's name. */
 } ObjFunction;
@@ -70,7 +70,7 @@ typedef struct {
  * @param Value*: pointer to the first argument on the stack.
  * @returns Value: the result value.
  */
-typedef Value (*NativeFn)(uint32_t argCount, Value* args);
+typedef Value (*NativeFn)(uint8_t argCount, Value* args);
 
 /**
  * Native function representation.
@@ -87,15 +87,15 @@ typedef struct {
  */
 struct ObjString {
 	Object object;
-	int32_t length;
+	uint32_t length;
 	char* chars;
 	uint32_t hash;
 };
 
 ObjFunction* newFunction(void);
 ObjNative* newNative(NativeFn function);
-ObjString* takeString(char* chars, int32_t length);
-ObjString* copyString(const char* chars, int32_t length);
+ObjString* takeString(char* chars, uint32_t length);
+ObjString* copyString(const char* chars, uint32_t length);
 void printObject(Value value);
 
 /** Safety check before downcasting from 'Object*' to one of the child object type.
